retries = 5

+++ download.sh
@@ -18,27 +18,31 @@ function increaseCounter {
     itc=`head -n1 it`
     let "itc+=1"
     echo itc > it
-
 }
 echo "I'm going to download ${1}"
 
 while true; do
+    if [[ $retries -le 0 ]]; then
+        echo "[$1] Sadly but it seems i already did everything i could. Terminating script with code 1 :("
+        exit 1
+    fi
     #Downloading Artifact
-    echo "Starting the download"
+    echo "[$1] Starting the download"
     if [[ $(aria2c ${ariaConfiguration} $baseURL/${1}) == 0 ]]; then
-        echo "Download finished successfully. Status code 0. Let me try to identify what it was"
+        echo "[$1] Download finished successfully. Status code 0. Let me try to identify what it was"
         #Identifying Artifact
         identifyArtifact ${1}
         if [[ $aBuildNumber == $buildNumberHosted ]]; then
-            echo "Great. That was what we expected. Storing the file"
+            echo "[$1] Great. That was what we expected. Storing the file"
             mv ${1} ${artifactsStorage}/${aVersion}-${aBuildNumber}-${1}
             linkArtifact ${aVersion}-${aBuildNumber}-${1} ${aBuildNumber}
             break
         else
-            echo "Unrecognized or unnecessary stuff. Burning it down and trying again"
+            echo "[$1] Unrecognized or unnecessary stuff. Burning it down and trying again"
             rm -rf ${1}
         fi
     fi
+    let "retries -= 1"
 done
 
 echo "${1} just have been downloaded successfully. All tests passed. Terminating the script with code 0"
